Refactor to put sets and challenges into correct modules
